Personalization for my teleirc bot.
I forked the project so I can keep all of my changes in my git profile and share with members of our group. Nothing is this fork is meant to be a proposal to the original project. At this time no change to base code, only edits to make our bot work with our IRC and Group.